span use int32

+++ datasets/file_util.py
@@ -36,8 +36,8 @@ def get_record_parser(options):
         qst_vocab_ids = tf.reshape(tf.decode_raw(features["qst_vocab_ids"], tf.int32), [options.max_qst_length])
         ctx_pos_ids = tf.reshape(tf.decode_raw(features["ctx_pos_ids"], tf.int32), [options.max_ctx_length])
         ctx_ner_ids = tf.reshape(tf.decode_raw(features["ctx_ner_ids"], tf.int32), [options.max_ctx_length])
-        span = features["span"]
-        question_id = features['question_id']
+        span = tf.cast(features["span"], tf.int32)
+        question_id = tf.cast(features['question_id'], tf.int32)
         qst_pos_ids = tf.reshape(tf.decode_raw(features['qst_pos_ids'], tf.int32), [options.max_qst_length])
         qst_ner_ids = tf.reshape(tf.decode_raw(features['qst_ner_ids'], tf.int32), [options.max_qst_length])
         ctx_in_qst = tf.reshape(tf.decode_raw(features['ctx_in_qst'], tf.int32), [options.max_ctx_length])
